test: narrow the load allowlist, add a seed corpus, and wire replay into fuzz.ps1
ArgumentException comes out of the load allowlist. An argument exception names a
parameter of a method the caller never invoked -- id, index, number -- so it can
only be an internal precondition escaping, never a considered refusal of a file.
Tolerating it hid a whole family: a dangling sheet relationship id, a cell value
that overflows a double, and a style index past the end of the table all
classified as rejections, and the run reported clean. XLibur says "this file is
wrong" with PartStructureException.

FormatException and OverflowException are still tolerated and probably should
not be, for the same reason; that is left until this list stops producing new
findings, and noted in the code so it is a decision rather than an oversight.

fuzz.ps1 grows -Replay and learns the workbook-structured target. Seeds now come
from a committed corpus under XLibur.Fuzz/corpus rather than being written
inline, so the starting point no longer exists only on the machine that ran the
script before. That corpus carries its own .gitattributes: the repository checks
source out as CRLF, which would rewrite a seed byte for byte.

Frame selection moves into StackSummary, shared by the pipeline and by replay.
When each had its own copy they disagreed, and the one used for findings blamed
System.Linq.ThrowHelper for a defect that was in the style decoder.

Fixes a harness bug that cost a false finding: the input stream was scoped to
the load, but XLWorkbook keeps a reference to it and rewinds it during SaveAs,
so an ObjectDisposedException came out of the write path looking exactly like a
write-path defect.
