+++ XLibur.Fuzz/corpus/address/absolute.txt
@@ -1,0 +1,1 @@
+$A$1
