+++ XLibur.Fuzz/corpus/address/qualified-range.txt
@@ -1,0 +1,1 @@
+Sheet1!$A$1:$C$10
