+++ XLibur.Fuzz/corpus/formula/array.txt
@@ -1,0 +1,1 @@
+SUMPRODUCT(B1:B3,B1:B3)
