+++ XLibur.Fuzz/corpus/formula/if.txt
@@ -1,0 +1,1 @@
+IF(A1>0,"yes","no")
