+++ XLibur.Fuzz/corpus/formula/sum.txt
@@ -1,0 +1,1 @@
+SUM(1,2)
