+++ XLibur.Fuzz/corpus/formula/vlookup.txt
@@ -1,0 +1,1 @@
+VLOOKUP(A1,B1:B3,1,FALSE)
